+++ src/integrations/svg/parse.rs
@@ -25,7 +25,7 @@ pub fn load_svg_from_bytes(bytes: &[u8]) -> Result<VelloSvg, VectorLoaderError> 
         local_transform_center: {
             let mut transform = Transform::default();
             transform.translation.x = width / 2.0;
-            transform.translation.y = -height / 2.0;
+            transform.translation.y = height / 2.0;
             transform
         },
         width,