+++ src/render/extract.rs
@@ -6,7 +6,7 @@ use bevy::{
     },
 };
 
-use super::{VelloEntityCountData, VelloFrameProfileData};
+use super::{SkipScaling, VelloEntityCountData, VelloFrameProfileData};
 use crate::prelude::*;
 
 #[derive(SystemSet, Debug, Hash, PartialEq, Eq, Clone)]
@@ -22,6 +22,8 @@ pub struct ExtractedVelloScene {
     pub scene: VelloScene,
     pub transform: GlobalTransform,
     pub ui_node: Option<ComputedNode>,
+    pub screen_space: Option<VelloScreenSpace>,
+    pub skip_scaling: Option<SkipScaling>,
 }
 
 pub fn extract_scenes(
@@ -39,6 +41,8 @@ pub fn extract_scenes(
                 &InheritedVisibility,
                 Option<&ComputedNode>,
                 Option<&RenderLayers>,
+                Option<&VelloScreenSpace>,
+                Option<&SkipScaling>,
             ),
             Without<SkipEncoding>,
         >,
@@ -51,8 +55,16 @@ pub fn extract_scenes(
     let mut views: Vec<_> = query_views.iter().collect();
     views.sort_unstable_by_key(|(camera, _)| camera.order);
 
-    for (scene, transform, view_visibility, inherited_visibility, ui_node, render_layers) in
-        query_scenes.iter()
+    for (
+        scene,
+        transform,
+        view_visibility,
+        inherited_visibility,
+        ui_node,
+        render_layers,
+        screen_space,
+        skip_scaling,
+    ) in query_scenes.iter()
     {
         // Skip if visibility conditions are not met
         if !view_visibility.get() || !inherited_visibility.get() {
@@ -69,6 +81,8 @@ pub fn extract_scenes(
                     transform: *transform,
                     scene: scene.clone(),
                     ui_node: ui_node.cloned(),
+                    screen_space: screen_space.cloned(),
+                    skip_scaling: skip_scaling.cloned(),
                 })
                 .insert(TemporaryRenderEntity);
             n_scenes += 1;